defend mana recover access

+++ New Unity Project/Assets/ManaCosuming.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class ManaCosuming : MonoBehaviour
 {
-    private int FullHP;
+    private float FullHP;
     // Start is called before the first frame update
     void Start()
     {
@@ -15,7 +15,10 @@ public class ManaCosuming : MonoBehaviour
     void Update()
     {
         this.transform.localPosition = new Vector3(320 * ((float)GameManager.PlayerMana / (float)FullHP) - 320, 0, 0);
+        
     }
+
+    
 
 }
